[do-not-merge] add before callback on ActionControlFilter
This PR explores one option to add more customized call back. Now
in `ActionControlFilter`, we hard-coded the behavior for the
predefined backends `RedisCached` and `RedisHammerspaced`. What to
do if we want to we want to support more specific behavior for the
backends defined elsewhere (like the one I'm working on)

This PR makes it possbile to run customized actions on before filter,
however the solution looks a bit hacky to me. Alternatively, I can
copy `action_controller` and `action_controller_filter` to my own
repo and modify accordingly.